+++ src/maruko_output.c
@@ -42,6 +42,8 @@ int maruko_output_init(MarukoOutput *output, const VencOutputUri *uri,
 	output->allow_unix_encoder_stall = allow_unix_encoder_stall ? 1 : 0;
 	output->send_errors = 0;
 	output->trunc_warned = 0;
+	output->bad_au_drops = 0;
+	output->low_water_slots = 0;
 	memset(&output->send_queue, 0, sizeof(output->send_queue));
 	__atomic_store_n(&output->socket_drops, 0, __ATOMIC_RELAXED);
 	__atomic_store_n(&output->socket_writes, 0, __ATOMIC_RELAXED);
@@ -76,6 +78,8 @@ int maruko_output_init_shm(MarukoOutput *output, const char *shm_name)
 	output->allow_unix_encoder_stall = 0;
 	output->send_errors = 0;
 	output->trunc_warned = 0;
+	output->bad_au_drops = 0;
+	output->low_water_slots = 0;
 	memset(&output->send_queue, 0, sizeof(output->send_queue));
 	__atomic_store_n(&output->socket_drops, 0, __ATOMIC_RELAXED);
 	__atomic_store_n(&output->socket_writes, 0, __ATOMIC_RELAXED);
@@ -115,6 +119,8 @@ int maruko_output_init_frame_shm(MarukoOutput *output, const char *shm_name)
 	output->allow_unix_encoder_stall = 0;
 	output->send_errors = 0;
 	output->trunc_warned = 0;
+	output->bad_au_drops = 0;
+	output->low_water_slots = 0;
 	memset(&output->send_queue, 0, sizeof(output->send_queue));
 	__atomic_store_n(&output->socket_drops, 0, __ATOMIC_RELAXED);
 	__atomic_store_n(&output->socket_writes, 0, __ATOMIC_RELAXED);